Code reorganization [WIP]
